Rename 'studio update' subcommand to 'studio upsert'
