+++ cmd/tools/studio/upsert/root.go
@@ -1,17 +1,17 @@
-package update
+package upsert
 
 import (
 	"github.com/spf13/cobra"
 )
 
 var RootCMD = &cobra.Command{
-	Use:   "update",
-	Short: "Update Studio resources",
+	Use:   "upsert",
+	Short: "Upsert Studio resources",
 	Long:  ``,
 }
 
 func init() {
 	RootCMD.AddCommand(applicationCMD)
-	RootCMD.AddCommand(studioCMD)
 	RootCMD.AddCommand(terraformCMD)
+	RootCMD.AddCommand(studioCMD)
 }
